requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==22.12.0
click==8.1.3
colorama==0.4.6
coverage==7.0.5
Django==4.1.5
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.25.1
exceptiongroup==1.1.0
factory-boy==3.2.1
Faker==16.6.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
platformdirs==2.6.2
pluggy==1.0.0
Pygments==2.14.0
pyrsistent==0.19.3
pytest==7.2.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
